+++ src/extras-karchive/libkarchivedirectory.zig
@@ -312,7 +312,7 @@ pub const karchivedirectory = struct {
     ///
     /// ` self: QtC.KArchiveDirectory `
     ///
-    pub fn Permissions(self: ?*anyopaque) u16 {
+    pub fn Permissions(self: ?*anyopaque) u32 {
         return @intCast(qtc.KArchiveEntry_Permissions(@ptrCast(self)));
     }
