+++ src/extras-karchive/libkarchivefile.zig
@@ -309,7 +309,7 @@ pub const karchivefile = struct {
     ///
     /// ` self: QtC.KArchiveFile `
     ///
-    pub fn Permissions(self: ?*anyopaque) u16 {
+    pub fn Permissions(self: ?*anyopaque) u32 {
         return @intCast(qtc.KArchiveEntry_Permissions(@ptrCast(self)));
     }
